refactor(evaluations): TeacherEvaluations sous 300 lignes — éclatement complet (H1)
TeacherEvaluations.vue 1065 → 279 lignes. Les données/filtres/dérivés restent
dans useTeacherEvaluations (déjà mergé) et la carte dans EvaluationCard (mergée) ;
on les CONSOMME sans les modifier. Éclatement du reste :
- useTeacherEvaluationActions.js (234) : état de la modale + handlers d'actions
  (créer/éditer/voir résultats/synchroniser/publier/prévisualiser/supprimer),
  paramétré par { evaluationsLMS, loadEvaluationsLMS } pour ne pas toucher au
  composable de données existant.
- TeacherEvalFilters.vue (248) : carte de filtres + barre des expirées.
- TeacherEvalStats.vue (119) : grille de statistiques.
- TeacherEvalEmptyState.vue (85) : état vide.
- TeacherEvalCreateModal.vue (277) : modale de création de version en ligne.

CSS déplacé VERBATIM. Parité : les filtres restent en v-model (type des valeurs
préservé, identique à l'original) ; les règles @media de .filters-grid/.stats-grid
sont déplacées dans leurs composants (CSS scoped). Services/routes/payloads/
alert/confirm/console identiques. Tous fichiers ≤300.

Dette PRÉEXISTANTE conservée verbatim : règles @media .eval-info-grid/.eval-actions/
.btn-action/.online-stats (CSS mort depuis l'extraction d'EvaluationCard) gardées
dans l'orchestrateur ; le TODO "Rediriger vers page d'édition" inchangé.

Tests : 14 verts (composable d'actions + 4 montages). Build vert.
